Update basic CMP tests
